Progress decoupling Planemo's core from click & CLI interactions.
