+++ planemo/cli.py
@@ -1,21 +1,15 @@
 """The module describes a CLI framework extending ``click``."""
 import functools
-import logging.config
 import os
-import shutil
 import sys
-import traceback
 
 import click
-from six.moves.urllib.request import urlopen
 
 from planemo import __version__
+from planemo.context import configure_standard_planemo_logging, PlanemoContext
 from planemo.exit_codes import ExitCodeException
 from planemo.galaxy import profiles
-from .config import (
-    OptionSource,
-    read_global_config,
-)
+from .config import OptionSource
 from .io import error
 
 
@@ -28,107 +22,23 @@ COMMAND_ALIASES = {
 }
 
 
-class Context(object):
-    """Describe context of Planemo computation.
+class PlanemoCliContext(PlanemoContext):
+    """Describe context of Planemo CLI computation.
 
-    Handles cross cutting concerns for Planemo such as verbose log
-    tracking, the definition of the Planemo workspace (``~/.planemo``),
-    and the global configuraton defined in ``~/.planemo.yml``.
+    Extend PlanemoContext with operations for CLI concerns (exit code) and
+    for interacting with the click library.
     """
 
-    def __init__(self):
-        """Construct a Context object using execution environment."""
-        self.home = os.getcwd()
-        self._global_config = None
-        # Will be set by planemo CLI driver
-        self.verbose = False
-        self.planemo_config = None
-        self.planemo_directory = None
-        self.option_source = {}
-
-    def set_option_source(self, param_name, option_source, force=False):
-        """Specify how an option was set."""
-        if not force:
-            assert param_name not in self.option_source, "No option source for [%s]" % param_name
-        self.option_source[param_name] = option_source
-
-    def get_option_source(self, param_name):
-        """Return OptionSource value indicating how the option was set."""
-        assert param_name in self.option_source, "No option source for [%s]" % param_name
-        return self.option_source[param_name]
-
-    @property
-    def global_config(self):
-        """Read Planemo's global configuration.
-
-        As defined most simply by ~/.planemo.yml.
-        """
-        if self._global_config is None:
-            self._global_config = read_global_config(self.planemo_config)
-        return self._global_config or {}
-
-    def log(self, msg, *args):
-        """Log a message to stderr."""
-        if args:
-            msg %= args
-        click.echo(msg, file=sys.stderr)
-
-    def vlog(self, msg, *args, **kwds):
-        """Log a message to stderr only if verbose is enabled."""
-        if self.verbose:
-            self.log(msg, *args)
-            if kwds.get("exception", False):
-                traceback.print_exc(file=sys.stderr)
-
-    @property
-    def workspace(self):
-        """Create and return Planemo's workspace.
-
-        By default this will be ``~/.planemo``.
-        """
-        if not self.planemo_directory:
-            raise Exception("No planemo workspace defined.")
-        workspace = self.planemo_directory
-        return self._ensure_directory(workspace, "workspace")
-
-    @property
-    def galaxy_profiles_directory(self):
-        """Create a return a directory for storing Galaxy profiles."""
-        path = os.path.join(self.workspace, "profiles")
-        return self._ensure_directory(path, "Galaxy profiles")
-
-    def _ensure_directory(self, path, name):
-        if not os.path.exists(path):
-            os.makedirs(path)
-        if not os.path.isdir(path):
-            template = "Planemo %s directory [%s] unavailable."
-            message = template % (name, path)
-            raise Exception(message)
-        return path
+    def _log_message(self, message):
+        click.echo(message, file=sys.stderr)
 
     def exit(self, exit_code):
         """Exit planemo with the supplied exit code."""
         self.vlog("Exiting planemo with exit code [%d]" % exit_code)
         raise ExitCodeException(exit_code)
 
-    def cache_download(self, url, destination):
-        cache = os.path.join(self.workspace, "cache")
-        if not os.path.exists(cache):
-            os.makedirs(cache)
-        filename = os.path.basename(url)
-        cache_destination = os.path.join(cache, filename)
-        if not os.path.exists(cache_destination):
-            with urlopen(url) as fh:
-                content = fh.read()
-            if len(content) == 0:
-                raise Exception("Failed to download [%s]." % url)
-            with open(cache_destination, "wb") as f:
-                f.write(content)
 
-        shutil.copy(cache_destination, destination)
-
-
-pass_context = click.make_pass_decorator(Context, ensure=True)
+pass_context = click.make_pass_decorator(PlanemoCliContext, ensure=True)
 cmd_folder = os.path.abspath(os.path.join(os.path.dirname(__file__),
                                           'commands'))
 
@@ -262,71 +172,15 @@ def planemo(ctx, config, directory, verbose, configure_logging=True):
     """
     ctx.verbose = verbose
     if configure_logging:
-        logging_config = {
-            'version': 1,
-            'disable_existing_loggers': False,
-            'formatters': {
-                'verbose': {
-                    'format': '%(name)s %(levelname)s %(asctime)s: %(message)s'
-                },
-                'simple': {
-                    'format': '%(name)s %(levelname)s: %(message)s'
-                },
-            },
-            'handlers': {
-                'console': {
-                    'level': 'DEBUG',
-                    'class': 'logging.StreamHandler',
-                    'formatter': 'simple' if not verbose else 'verbose'
-                },
-            },
-            'loggers': {
-                # Suppress CWL is beta warning, for Planemo purposes - it is absolutely not.
-                'galaxy.tools.parser.factory': {
-                    'handlers': ['console'],
-                    'propagate': False,
-                    'level': 'ERROR' if not verbose else "DEBUG",
-                },
-                'galaxy.tools.deps.commands': {
-                    'handlers': ['console'],
-                    'propagate': False,
-                    'level': 'ERROR' if not verbose else "DEBUG",
-                },
-                'galaxy': {
-                    'handlers': ['console'],
-                    'propagate': False,
-                    'level': 'INFO' if not verbose else "DEBUG",
-                },
-                # @jmchilton
-                # I'm fixing up Planemo's lint functionality for CWL and I keep seeing this for the
-                # schema metadata stuff (e.g. in the workflows repo). "rdflib.term WARNING:
-                # http://schema.org/docs/!DOCTYPE html does not look like a valid URI, trying to
-                # serialize this will break.". I'm going to suppress this warning I think, or are the
-                # examples wrong and should declare their namespaces differently in some way?
-                # @mr-c
-                # That particular warning is worth suppressing. A PR to silence it permanently would be very welcome!
-                # https://github.com/RDFLib/rdflib/blob/master/rdflib/term.py#L225
-                'rdflib.term': {
-                    'handlers': ['console'],
-                    'propagate': False,
-                    'level': 'ERROR' if not verbose else "DEBUG",
-                }
-            },
-            'root': {
-                'handlers': ['console'],
-                'propagate': False,
-                'level': 'WARNING' if not verbose else "DEBUG",
-            }
-        }
-        logging.config.dictConfig(logging_config)
+        configure_standard_planemo_logging(verbose)
     ctx.planemo_config = os.path.expanduser(config)
     ctx.planemo_directory = os.path.expanduser(directory)
 
 
 __all__ = (
     "command_function",
-    "Context",
     "list_cmds",
     "name_to_command",
     "planemo",
+    "PlanemoCliContext",
 )
