+++ planemo/config.py
@@ -1,19 +1,20 @@
 """Module defines abstractions for configuring Planemo."""
 
 import os
+from enum import Enum
+from typing import Any, Dict
 
-import aenum
 import click
 import yaml
 
 PLANEMO_CONFIG_ENV_PROP = "PLANEMO_GLOBAL_CONFIG_PATH"
 DEFAULT_CONFIG = {
-}
+}  # type: Dict[str, Any]
 
 VALUE_UNSET = object()
 
 
-OptionSource = aenum.Enum(
+OptionSource = Enum(
     "OptionSource", 'cli profile global_config default'
 )
